raspberry_pi/requirements.txt
requests>=2.31.0
python-dotenv>=1.1.0
pygame>=2.5.0  # For audio playback in raspi_client.py
opencv-python>=4.8.0  # For video capture in raspi_stream.py
numpy>=1.24.0  # Used by opencv 